pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "news-manager"
version = "0.1.0"
description = "Fetch news sources, filter and summarize articles with Groq"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27,<1",
    "beautifulsoup4>=4.12,<5",
    "trafilatura>=1.9,<2",
    "lxml>=5,<6",
    "lxml-html-clean>=0.4,<1",
    "openai>=1.40,<2",
    "python-dotenv>=1.0,<2",
]

[project.optional-dependencies]
dev = ["pytest>=8,<9"]

[project.scripts]
news-manager = "news_manager.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["news_manager*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
